feat: Add AI Element Filter and Operation Tools - Add ai_element_filter: A universal tool for retrieving specific Revit element information - Add operate_element: A universal tool for performing operations on Revit elements These new components provide a standardized approach for filtering and manipulating Revit elements, improving code consistency and reducing duplication across the project.

+++ revit-mcp-commandset/Services/GetCurrentViewInfoEventHandler.cs
@@ -1,6 +1,6 @@
 ﻿using Autodesk.Revit.UI;
 using RevitMCPCommandSet.Models.Common;
-using revit_mcp_sdk.API.Interfaces;
+using RevitMCPSDK.API.Interfaces;
 
 namespace RevitMCPCommandSet.Services
 {
@@ -30,7 +30,7 @@ namespace RevitMCPCommandSet.Services
                 ResultInfo = new ViewInfo
                 {
 #if REVIT2024_OR_GREATER
-                    Id = activeView.Id.Value,
+                    Id = (int)activeView.Id.Value,
 #else
                     Id = activeView.Id.IntegerValue,
 #endif
